pyproject.toml
[tool.poetry]
name = "ppr-helm-imageupdater"
version = "0.1.0"
description = "Image updater for helm"
authors = ["Alex Parra <jparra@occ.com.mx>"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
bpython = "^0.22.1"
pylint = "^2.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
